pkp/pkp-lib#2461: Add section_ref as required attribute to articles; improve error recovery for native import when objects cannot be added to the database due to foreign key constraints

+++ plugins/importexport/native/NativeImportExportDeployment.inc.php
@@ -91,8 +91,10 @@ class NativeImportExportDeployment extends PKPNativeImportExportDeployment {
 				if (!empty($processedIssuesIds)) {
 					$issueDao = DAORegistry::getDAO('IssueDAO');
 					foreach ($processedIssuesIds as $issueId => $errorMessages) {
-						$issue = $issueDao->getById($issueId);
-						$issueDao->deleteObject($issue);
+						if ($issueId) {
+							$issue = $issueDao->getById($issueId);
+							$issueDao->deleteObject($issue);
+						}
 					}
 				}
 				break;
